+++ src/components/ui/animated-testimonials.tsx
@@ -58,7 +58,7 @@ const AnimatedCanopy = ({
 );
 
 const TestimonialCard = ({ testimonial }: { testimonial: Testimonial }) => (
-  <div className="group mx-2 flex h-32 w-80 shrink-0 cursor-pointer overflow-hidden rounded-xl border border-transparent bg-background p-3 transition-all hover:border-blue-400 hover:shadow-[0_0_10px_#60a5fa] dark:hover:border-blue-400">
+  <div className="group mx-2 flex h-32 w-80 shrink-0 cursor-pointer overflow-hidden rounded-xl border border-transparent  p-3 transition-all hover:border-blue-400 hover:shadow-[0_0_10px_#60a5fa] dark:hover:border-blue-400">
     <div className="flex items-start gap-3">
       <div className="relative h-12 w-12 shrink-0 overflow-hidden rounded-full border-2 border-gray-200 dark:border-gray-600">
         <img
